requirements.txt
aiohttp==3.8.6
aiosignal==1.3.1
asttokens==2.4.0
async-timeout==4.0.3
attrs==23.1.0
Authlib==1.2.1
blinker==1.6.2
build==0.10.0
CacheControl==0.13.1
certifi==2023.7.22
cffi==1.16.0
charset-normalizer==3.3.0
cleo==2.0.1
click==8.1.7
colorama==0.4.6
contourpy==1.1.1
crashtest==0.4.1
cryptography==41.0.4
cycler==0.12.1
distlib==0.3.7
dulwich==0.21.6
executing==2.0.0
filelock==3.12.4
Flask==3.0.0
fonttools==4.43.1
frozenlist==1.4.0
icecream==2.1.3
idna==3.4
importlib-metadata==6.8.0
installer==0.7.0
itsdangerous==2.1.2
jaraco.classes==3.3.0
Jinja2==3.1.2
jsonschema==4.17.3
keyring==24.2.0
kiwisolver==1.4.5
langdetect==1.0.9
MarkupSafe==2.1.3
matplotlib==3.8.0
more-itertools==10.1.0
msgpack==1.0.7
multidict==6.0.4
numpy==1.26.1
openai==0.28.1
packaging==23.2
pexpect==4.8.0
Pillow==10.1.0
pkginfo==1.9.6
platformdirs==3.11.0
poetry==1.6.1
poetry-core==1.7.0
poetry-plugin-export==1.5.0
ptyprocess==0.7.0
pycparser==2.21
Pygments==2.16.1
pyparsing==3.1.1
pyproject_hooks==1.0.0
pyrsistent==0.19.3
python-dateutil==2.8.2
python-dotenv==1.0.0
pywin32-ctypes==0.2.2
rapidfuzz==2.15.2
requests==2.31.0
requests-toolbelt==1.0.0
schedule==1.2.1
sentry-sdk==1.32.0
shellingham==1.5.3
six==1.16.0
slack-bolt==1.18.0
slack-sdk==3.23.0
tomli==2.0.1
tomlkit==0.12.1
tqdm==4.66.1
trove-classifiers==2023.10.18
urllib3==2.0.6
validators==0.22.0
virtualenv==20.24.5
weaviate-client==3.24.2
Werkzeug==3.0.0
yarl==1.9.2
zipp==3.17.0
